+++ CHANGELOG.md
@@ -46,3 +46,7 @@ on every vscode update a new version will be release
 ## [1.85.1008] - 01-19-2024
 * [UPDATE] added InfiniteGlobString class to rpresent the value at the base of InfiniteGlobStringArray or the includeGlobs option
 and added InfiniteGlobString.createFileIfNotFoundPath on that so that the extension can create the file if the related files does not exist
+
+## [1.85.1009] - 01-19-2024
+* [UPDATE] added InfiniteGlobString class to rpresent the value at the base of InfiniteGlobStringArray or the includeGlobs option
+and added InfiniteGlobString.createFileIfNotFoundPath on that so that the extension can create the file if the related files does not exist